refactor(kpilote): retours de revue carte France (schéma, cast, boundary, prop)
- Déplace schéma/type GeoJSON de src/api/geoJson.ts (plus d'API) vers
  src/assets/maps/geoJson.ts, co-localisé avec les cartes.
- src/assets/maps/index.ts : franceGeoJsonSchema.parse au lieu du cast
  (type sans cast + validation bon marché, geometry non parcourue).
- CarteFrance : prop frontieres requise (toujours fournie) ; suppression
  des branches conditionnelles.
- IndicateurResultatsTab : retire le boundary Suspense devenu redondant
  (referentiel/individus/remarquables déjà préfetchés par le loader, carte
  en import statique).
