+++ apps/kpilote-webapp/src/assets/maps/geoJson.ts
@@ -6,8 +6,9 @@ const geoJsonFeatureSchema = z.object({
   properties: z.object({ code: z.string(), nom: z.string() }),
 })
 
-// Contrat des cartes France. Validé à la génération (`pnpm maps:generate`) ;
-// les cartes sont ensuite importées statiquement (cf. `@/assets/maps`).
+// Contrat des cartes France. Validé à la génération (`pnpm maps:generate`) et à
+// l'import (cf. `index.ts`). `geometry: unknown` : on ne parcourt pas les
+// coordonnées, la validation reste bon marché.
 export const franceGeoJsonSchema = z.object({
   type: z.literal('FeatureCollection'),
   features: z.array(geoJsonFeatureSchema).readonly(),
